moved common code in lib
